pyproject.toml
[project]
name = "openai-import-memory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "graphiti-core>=0.18.9",
    "mem0ai>=0.1.116",
    "pyyaml>=6.0.2",
    "tqdm>=4.67.1",
]
